Removed Redundant Files and Moved Enum to Enum folder
