+++ SoftServeCarMobile/Environment.tsx
@@ -1,5 +1,6 @@
 export const environment = {
     url: "https://car-project.azurewebsites.net/"
+    //url: "http://192.168.0.105:5000/"
 };
 
 export const routes = {
