+++ SoftServeCarMobile/models/Notification.tsx
@@ -1,3 +1,5 @@
+import {NotificationType} from "../src/common/enums/NotificationType";
+
 export type Notification = null | {
     id: number;
     userId: number;
@@ -11,15 +13,3 @@ export type Notification = null | {
     userColor: string;
     notificationType: NotificationType;
 };
-export enum NotificationType {
-    PassengerApply = 1,
-    ApplicationApproval = 2,
-    JourneyCancellation = 3,
-    JourneyDetailsUpdate = 4,
-    JourneyInvitation = 5,
-    AcceptedInvitation = 6,
-    RejectedInvitation = 7,
-    PassengerWithdrawal = 8,
-    HRMarketingMessage = 9,
-    HRMarketingSurvey = 10
-}
