+++ SoftServeCarMobile/src/common/enums/NotificationType.ts
@@ -1,0 +1,12 @@
+export enum NotificationType {
+    PassengerApply = 1,
+    ApplicationApproval = 2,
+    JourneyCancellation = 3,
+    JourneyDetailsUpdate = 4,
+    JourneyInvitation = 5,
+    AcceptedInvitation = 6,
+    RejectedInvitation = 7,
+    PassengerWithdrawal = 8,
+    HRMarketingMessage = 9,
+    HRMarketingSurvey = 10
+}
